A bit of refactor
