+++ app/ante.go
@@ -1,27 +1,19 @@
 package app
 
 import (
-	"fmt"
-	"runtime/debug"
-
 	errorsmod "cosmossdk.io/errors"
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
-	"github.com/cosmos/cosmos-sdk/crypto/types/multisig"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	"github.com/cosmos/cosmos-sdk/x/auth/ante"
-	"github.com/cosmos/cosmos-sdk/x/auth/types"
 	conntypes "github.com/cosmos/ibc-go/v6/modules/core/03-connection/types"
 	ibcante "github.com/cosmos/ibc-go/v6/modules/core/ante"
 	ibckeeper "github.com/cosmos/ibc-go/v6/modules/core/keeper"
 	"github.com/dymensionxyz/dymension-rdk/x/gasless"
 	gaslesskeeper "github.com/dymensionxyz/dymension-rdk/x/gasless/keeper"
 	cosmosante "github.com/evmos/evmos/v12/app/ante/cosmos"
-	"github.com/evmos/evmos/v12/crypto/ethsecp256k1"
-	tmlog "github.com/tendermint/tendermint/libs/log"
 )
 
 // HandlerOptions are the options required for constructing a default SDK AnteHandler.
@@ -34,23 +26,16 @@ type HandlerOptions struct {
 	GaslessKeeper     gaslesskeeper.Keeper
 }
 
-func NewAnteHandler(options HandlerOptions) (sdk.AnteHandler, error) {
+// NewAnteHandler returns an AnteHandler that checks and increments sequence
+// numbers, checks signatures & account numbers, and deducts fees from the first
+// signer.
+func NewAnteHandler(options HandlerOptions) sdk.AnteHandler {
 	if err := options.validate(); err != nil {
-		return nil, fmt.Errorf("options validate: %w", err)
+		panic(err)
 	}
 
-	return func(ctx sdk.Context, tx sdk.Tx, sim bool) (_ sdk.Context, err error) {
-		defer Recover(ctx.Logger(), &err)
-
-		return cosmosHandler(options)(ctx, tx, sim)
-	}, nil
-}
-
-func cosmosHandler(options HandlerOptions) sdk.AnteHandler {
 	return sdk.ChainAnteDecorators(getAnteDecorators(options)...)
 }
-
-type SigGasConsumer func(meter sdk.GasMeter, sig signing.SignatureV2, params types.Params) error
 
 func getAnteDecorators(options HandlerOptions) []sdk.AnteDecorator {
 	anteDecorators := []sdk.AnteDecorator{
@@ -73,7 +58,7 @@ func getAnteDecorators(options HandlerOptions) []sdk.AnteDecorator {
 		NewBypassIBCFeeDecorator(gasless.NewDeductFeeDecorator(options.AccountKeeper, options.BankKeeper, options.FeegrantKeeper, options.TxFeeChecker, options.GaslessKeeper)),
 		ante.NewSetPubKeyDecorator(options.AccountKeeper), // SetPubKeyDecorator must be called before all signature verification decorators
 		ante.NewValidateSigCountDecorator(options.AccountKeeper),
-		ante.NewSigGasConsumeDecorator(options.AccountKeeper, defaultSigVerificationGasConsumer),
+		ante.NewSigGasConsumeDecorator(options.AccountKeeper, options.SigGasConsumer),
 		NewSigCheckDecorator(options.AccountKeeper.(accountKeeper), options.SignModeHandler),
 		ante.NewIncrementSequenceDecorator(options.AccountKeeper),
 	}
@@ -81,59 +66,6 @@ func getAnteDecorators(options HandlerOptions) []sdk.AnteDecorator {
 	anteDecorators = append(anteDecorators, ibcante.NewRedundantRelayDecorator(options.IBCKeeper))
 
 	return anteDecorators
-}
-
-const (
-	secp256k1VerifyCost uint64 = 21000
-)
-
-// TODO: check with zero fee relayer
-// Copied from github.com/evmos/ethermint
-func defaultSigVerificationGasConsumer(meter sdk.GasMeter, sig signing.SignatureV2, params types.Params) error {
-	pubkey := sig.PubKey
-	switch pubkey := pubkey.(type) {
-	case *ethsecp256k1.PubKey:
-		meter.ConsumeGas(secp256k1VerifyCost, "ante verify: eth_secp256k1")
-		return nil
-
-	case multisig.PubKey:
-		// Multisig keys
-		multisignature, ok := sig.Data.(*signing.MultiSignatureData)
-		if !ok {
-			return fmt.Errorf("expected %T, got, %T", &signing.MultiSignatureData{}, sig.Data)
-		}
-		return consumeMultisignatureVerificationGas(meter, multisignature, pubkey, params, sig.Sequence)
-
-	default:
-		return ante.DefaultSigVerificationGasConsumer(meter, sig, params)
-	}
-}
-
-// Copied from github.com/evmos/ethermint
-func consumeMultisignatureVerificationGas(
-	meter sdk.GasMeter, sig *signing.MultiSignatureData, pubkey multisig.PubKey,
-	params types.Params, accSeq uint64,
-) error {
-	size := sig.BitArray.Count()
-	sigIndex := 0
-
-	for i := 0; i < size; i++ {
-		if !sig.BitArray.GetIndex(i) {
-			continue
-		}
-		sigV2 := signing.SignatureV2{
-			PubKey:   pubkey.GetPubKeys()[i],
-			Data:     sig.Signatures[sigIndex],
-			Sequence: accSeq,
-		}
-		err := defaultSigVerificationGasConsumer(meter, sigV2, params)
-		if err != nil {
-			return err
-		}
-		sigIndex++
-	}
-
-	return nil
 }
 
 func (o HandlerOptions) validate() error {
@@ -154,23 +86,4 @@ func (o HandlerOptions) validate() error {
 		return errorsmod.Wrap(sdkerrors.ErrLogic, "wasm config is required for ante builder")
 	}
 	return nil
-}
-
-func Recover(logger tmlog.Logger, err *error) {
-	if r := recover(); r != nil {
-		*err = errorsmod.Wrapf(sdkerrors.ErrPanic, "%v", r)
-
-		if e, ok := r.(error); ok {
-			logger.Error(
-				"ante handler panicked",
-				"error", e,
-				"stack trace", string(debug.Stack()),
-			)
-		} else {
-			logger.Error(
-				"ante handler panicked",
-				"recover", fmt.Sprintf("%v", r),
-			)
-		}
-	}
 }
